Introduce multi-key retrieval interface
This is the the interface which will be used to support key rotation.
The idea is that the keys are returned sorted by creation time, from
newest to oldest, so that the clients can try decrypting with all of
them. Since we do not store timestamps in AcraStructs, we have no direct
idea about which historical key has been used, so we have no other way
than trying them all.

We keep it in a new interface rather than adding to KeyStore directly
because we want to control which tools are rotation-aware, and which
are not. Obviously, direct users for filesystem/KeyStore will also be
able to access this interface.

The implementation is a dummy stub for now since we don't actually store
older keys. However, it is semantically correct.