Read historical keys on retrieval
We are going to keep historical keys in a subdirectory of the key store,
like this:

    ${ZONE_ID}_zone                - current key value, like before
    ${ZONE_ID}_zone.old/           - historical key directory
    |- 2020-02-09T07:17:23.76463
    `- 2020-02-12T13:40:11.674983

Retrieval is a bit convoluted because of all path juggling that we need
to perform, but it's not that hard. Note how errors are handled.

We provide a function to generate timestamped names for historical keys.
It is not used right now but we'll need it in rotation implementation.