Keep historical key data during updates
Introduce helper functions for writing key files and use them. Now it's
easier to check that we use correct permissions when saving a key *and*
we don't have to go on a safari for ioutil.WriteFile() calls when
updating key rotation code.

Instead of simply writing over existing key data, move the old key file
into historical storage (creating it if necessary) and store the new
content in the old place.

This operation is somewhat involved to ensure that the final update
happens atomically in the filesystem. That is, we never lose key data:
either the old key data is in place, or it has been backed up *and* new
key data is available.