Replace bytes.Buffer with plain append to avoid panics
Acra needs to support go 1.11 and that version is susceptible to a weird
false positive in cgo [1], triggered by buffers built by bytes.Buffer
which cause a panic in cgo pointer verification. We use GoThemis for
encryption and it uses cgo, and we use bytes.Buffer to construct context
buffers, resulting in a panic with go 1.11.5.

[1]: https://github.com/golang/go/issues/14210

Avoid this issue by replacing bytes.Buffer with plain slices and
appends. This arguably makes the code a bit easier to read too.