Replace custom utils with unified path separators
Instead of adding some new utilities for the users to deal with, accept
either forward or backward slashes as key path separators. We'll convert
them into OS separators as needed but otherwise they are treated as
identical.

Now the users will be able to use filepath.Join() to construct key ring
paths if they wish to. They can also manually join with "/", or hardcode
paths with "/".