acra-rotate: use RotateStorageKeyStore
This tool needs to manipulate storage keys and read them back to perform
rotation. Introduce a specialized interface with these traits.

Note that we can actually reuse existing interfaces and avoid
implementing a separate "rotateKeystore". Basically, all we need to do
is to use SaveDataEncryptionKeys() instead of SaveClientIDKeypair().
Then we can use the regular filesystem.KeyStore for this job.