acra-server: use ServerKeyStore, DecryptionKeyStore
Introduce a specialized key store for acra-server which includes almost
everything. acrs-server provides management API so it needs to have
access to all keystore abilities.

Most of the internal processing can be performed with read-only access
to keys. Introduce a DecryptionKeyStore interface which defines this.
It can also be used by acra-translator.