Overhaul signature interface
Provide extensible interface to define signature algorithms. It leaks
implementation detail of ASN.1 but that should not be an issue in
practice. However, with this it is possible to teach signature.Notary
to support arbitrary signature algorithms.

We keep the KeyStoreSuite entity but it's reclassified into a struct
which acts as a container for cryptography we'll be using. We also keep
convenience NewSCellSuite() function to produce a default cryptosuite.

This also allows to avoid copying Secure Cell encryptor and just simply
reuse existing implementation from keystoreV1.