Initial Key Store v2 (#352)
* Initial keystore V2

Add initial version of keystore V2 (code name: key rings). This is
cleaned up, rewritten, and expanded prototype code which implements
more complex data model for stored keys, aimed at better compliance
with NIST SP 800-57 recommendations.

Now "key store" stores not simply 'keys' but 'key rings' that contain
multiple key versions in them, as well as metadata describing keys:
their purpose, validity range (cryptoperiod), and current state.

Key ring descriptions are stored as structured data described by ASN.1
grammar and serialized using DER. This allows us to sign resulting key
ring data to ensure that it is not modified by a third-party.

This commit includes a "filesystem keystore" abstraction similar to the
one we have right now. It provides an interface to a file-base key store
with an additional intermediate interface allowing to plug arbitrary
backends (may be useful to support key-value stores, for example).

Note that the implementation is somewhat complicated because we need to
update key ring data atomically and prevent possible data loss, either
due to power failures, application failures, or concurrent modification.

Also note that this is only an initial version, so it currently lacks
some important features:

  - The key store is not integrated with the rest of the services.
    It does not implement the KeyStore V1 interface right now.
    Implementing high-level key store interface on top of this one
    is the next step.

  - Key directory integrity is not verified. Current implementation
    verifies that key ring data is intact, but does not verify that
    key rings themselves are not removed or entirely replaced. The
    data model includes a 'key directory' entity to track this, but
    they are not implemented currently.

* Fix go 1.11 compatibility

We still need to support go 1.11. Unfortunately, strings.ReplaceAll()
is available only since go 1.12. Use a regular Replace() instead.

* Compare times with Equal()

We'd better compare them with Equal() instead of using == because it
does not work on all systems (like, Linux running in a container with
a different time zone).

* Replace bytes.Buffer with plain append to avoid panics

Acra needs to support go 1.11 and that version is susceptible to a weird
false positive in cgo [1], triggered by buffers built by bytes.Buffer
which cause a panic in cgo pointer verification. We use GoThemis for
encryption and it uses cgo, and we use bytes.Buffer to construct context
buffers, resulting in a panic with go 1.11.5.

[1]: https://github.com/golang/go/issues/14210

Avoid this issue by replacing bytes.Buffer with plain slices and
appends. This arguably makes the code a bit easier to read too.

* Make Unmarshal() standalone functions

This is more in-line with other serialization packages in Go, like
encoding/asn1 itself.

* Replace custom utils with unified path separators

Instead of adding some new utilities for the users to deal with, accept
either forward or backward slashes as key path separators. We'll convert
them into OS separators as needed but otherwise they are treated as
identical.

Now the users will be able to use filepath.Join() to construct key ring
paths if they wish to. They can also manually join with "/", or hardcode
paths with "/".

* Move key access API to KeyRing directly

Instead of this indirection via a separate "Key" interface, access keys
directly from KeyRings. These should not be use cases where we need to
operate a key separately from its key ring context. Thus move all Key
methods into KeyRing interface. This means that we have to pass seqnums
explicitly. However, it also greatly simplifies implementation.

* Overhaul signature interface

Provide extensible interface to define signature algorithms. It leaks
implementation detail of ASN.1 but that should not be an issue in
practice. However, with this it is possible to teach signature.Notary
to support arbitrary signature algorithms.

We keep the KeyStoreSuite entity but it's reclassified into a struct
which acts as a container for cryptography we'll be using. We also keep
convenience NewSCellSuite() function to produce a default cryptosuite.

This also allows to avoid copying Secure Cell encryptor and just simply
reuse existing implementation from keystoreV1.

* Prevent accidental collisions in signatures

Insert a separator between "context" and "data" when computing HMAC
signatures. This ensures that it's not possible to achieve collision
by shifting some bytes between context and data:

    Sign("hello ", "world!") != Sign("hell", "o world!)