Split KeyStore interface (#353)
* Split KeyStore interface into more subinterfaces

Currently we have separate SecureSessionKeyStore, PublicKeyStore and
PrivateKeyStore. Let's make more interfaces which segregate keys by
their purpose and use. This makes it easier to delineate which services
have need to known which keys.

* acra-authmanager: use WebConfigKeyStore

authmanager basically needs to access only WebConfig's symmetric key.
It's okay to provide it with this interface.

* acra-addzone: use StorageKeyCreation

This tool needs to be able to manipulate storage keys in order to add
new zone keys. It does not need to do anything else.

* acra-poisonrecordmaker: use PoisonKeyStore

This tool can be limited to using its specialized PoisonKeyStore
interface which provides access to poison record key alone.

* acra-rotate: use RotateStorageKeyStore

This tool needs to manipulate storage keys and read them back to perform
rotation. Introduce a specialized interface with these traits.

Note that we can actually reuse existing interfaces and avoid
implementing a separate "rotateKeystore". Basically, all we need to do
is to use SaveDataEncryptionKeys() instead of SaveClientIDKeypair().
Then we can use the regular filesystem.KeyStore for this job.

* acra-translator: use TranslationKeyStore

This tool needs to get access to storage keys, transport keys, and the
poison record key. Introduce a specialized key store interface for this.

* acra-connector: use TransportKeyStore

It's mostly the same as SecureSessionKeyStore but has one additional
method in it. I wonder if we can merge them... However, there are
several places which specifically use SecureSessionKeyStore and we
don't want to change its definition.

* acra-keymaker: use KeyMaking

Introduce a specialized key store access interface which collects all
key store update interface. keymaker tool obviously needs to be able
to write all those keys.

* acra-server: use ServerKeyStore, DecryptionKeyStore

Introduce a specialized key store for acra-server which includes almost
everything. acrs-server provides management API so it needs to have
access to all keystore abilities.

Most of the internal processing can be performed with read-only access
to keys. Introduce a DecryptionKeyStore interface which defines this.
It can also be used by acra-translator.

* Drop original KeyStore interface

Now that all services have been updated to use their own specialized
interfaces, there is no need for a big "KeyStore" interface which
does everything. Drop it. We're going to reuse the name KeyStore in v2
for a different, slightly more low-level interface.