Log formatting hooks
Provide extensibility hooks for log formatters which allow to inspect
and modify formatted entries.

Use logging.CustomizeLoggingWithHooks() to configure a logger with
additional hooks. Installed hooks will be executed for each entry:

  - WillFormat() is called before the entry is formatted. You may
    inspect its fields, add new extra fields, etc.

  - DidFormat() is called after the entry is formatter. The original
    entry as well as a formatter buffer are passed to the hooks. They
    may inspect the fields once again and process serialized buffer.

Hooks may return an error, in which case formatting is stopped and this
error is returned to the caller.