Key export API for key store v1
Introduce key export API for filesystem.KeyStore. This API can be used
to enumerate keys stored in the key store. It will be particularly
useful for converting keystore v1 --> v2.

There is no better way for inspecting the hierarchy of keys other than
walking the entire key directory and matching private keys with their
public counterparts. Therefore we have to first enumerate all available
keys and then proceed with their export.

Key store v1 does not provide much metadata about the keys, so their
purpose and client/zone ID is all that we've got (besides key material).

Key metadata is not stored explicitly, but rather is encoded in the key
filename. We have to parse the filename to extract it. Key purpose is
encoded as string for this reason. It enables extensibility as Acra EE
supports more keys that Acra CE.

Acra CE is supported to use EnumerateExportedKeys() to export keys while
Acra EE will use EnumerateExportedKeysByClass with a derived classifier.
The classifier may use NewExported*Key() functions to describe keys.

Provide additional utilities for exporting symmetric keys that are
stored either encrypted or as plaintext. Currently Acra CE keeps just
authentication symmetric key which is stored as plaintext. However,
Acra EE keeps many of its symmetric keys encrypted.

Since storing keys in plaintext is definitely not a norm, provide a
separate API for dealing with them, giving shorter names to the
encrypted key API.

We kind of abuse the "ID" field for symmetric keys to store encryption
context there. It's probably okay for now since we are not going to
access it directly.