Key import API for key store v2
Add low-level method ImportKeyFileV1() which allows key store v2 to
import data exported from key store v1.

This method imports only a single key but usually you would like to
import an entire key store. It leaves transactionality to the caller.
That is, we must ensure that the key store is in consistent state
(ideally, unchanged) if importing process fails midway. It's not easy
to rollback changes so we'll introduce restrictions on importing.

Note that KeyFileImportV1 is also an interface which allows Acra EE
to override and refine importing process. Normally it would handle
extensions and then delegate keys with unknown purpose to Acra CE.