Key store version marker
Leave a special "version" file in filesystem-based key stores so that
it's easier to identify their expected content and distinguish new
key stores from older ones.

The file is created when a new key store is initialized. After that
it is never modified, but checked to verify that it keeps expected
value whenever we open an existing key store.

There is also an IsKeyDirectory() utility to check the version file
without opening the key store. This is not a complete check, intended
to guess what we're dealing with in absence of specific information.