Detect key store version by content
Instead of assuming the key store to be v1 unless --keystore is given,
look into the key directory and check whether it looks like a key store.
It seems to be v2 then open it as v2, otherwise assume v1. This applies
to acra-keymaker too which will still create key store v1 by default.

The user can explicitly request key store v2 with --keystore=v2.
Normally this is required only when creating the first key, after that
the key store format should be detected automatically.