acra-migrate-keys tool
Create a tool for migrating keys between key store formats. Currently it
supports only v1 -> v2 migration.

With default settings it is enough to simply run "acra-migrate-keys" to
migrate keys from ".acrakeys" (v1) into ".acrakeys.migrated" (v2). After
successful migration the administrator has to manually reconfigure Acra
services to use the new key store. It is possible to explicitly specify
migration details via --src... and --dst...  options if necessary.

There is a --dry_run option which performs migration in memory, without
writing results on disk. You can use it to test that everything is in
place, the keys can be read and understood.

By default the migration tool will refuse to write output into already
existing key store. This is because the migration process is not
transactional and idempotent. Migrating the same key store twice will
result in duplicate keys, and failure to migrate some keys will still
migrate other successful keys. This behavior can be overriden with
--force option, allowing the migration to proceed.

Note that master keys are expected in environment variables, as usual.
You will need to set all variables for successful migration:

  - ACRA_MASTER_KEY
  - ACRA_MASTER_ENCRYPTION_KEY
  - ACRA_MASTER_SIGNATURE_KEY

Most of the functions and data structures of the migration tool are
exported. This is an experimental attempt to reduce possible code
duplication in Acra EE migration tool. Hopefully it will be able to
reuse some of this code.