Use deferred error pattern correctly
We have been using a "deferred error" pattern in some places. Its idea
is to perform some cleanup action that might fail and return that error,
unless the whole operation failed due to some other error, in which case
the cleanup is still performed, but the earlier error is returned.

The correct way to do this is like this:

    func fallibleTransaction() (err error) {    // 1
            defer func() {
                    err2 := cleanup()
                    if err == nil {
                            err = err2          // 2
                    }
            }()
            err = operation()
            if err != nil {
                    return err
            }
            return nil                          // 3
    }

Note that you have to use return variables (1) in order to have it
assignable from defer, return the cleanup error only if the original
operation has not failed (2). You can use normal returns elsewhere (3),
defer will be able to fix up the return value if necessary, even if
the code explicitly says "return nil".

Fix up existing code using this pattern to use it correctly.