Dry run acra-migrate-keys in configuration tests
Configuration test currently run all available tools. Make sure that
acra-migrate-keys is a) built, b) run with --dry_run and correct path
to existing key store v1 directory.

This is not a proper and final way to test it, just make sure the tests
don't break. Later we will need to be more accurate with key store
versioning and paths.