acra-migrate-keys tool (#356)
* Key export API for key store v1

Introduce key export API for filesystem.KeyStore. This API can be used
to enumerate keys stored in the key store. It will be particularly
useful for converting keystore v1 --> v2.

There is no better way for inspecting the hierarchy of keys other than
walking the entire key directory and matching private keys with their
public counterparts. Therefore we have to first enumerate all available
keys and then proceed with their export.

Key store v1 does not provide much metadata about the keys, so their
purpose and client/zone ID is all that we've got (besides key material).

Key metadata is not stored explicitly, but rather is encoded in the key
filename. We have to parse the filename to extract it. Key purpose is
encoded as string for this reason. It enables extensibility as Acra EE
supports more keys that Acra CE.

Acra CE is supported to use EnumerateExportedKeys() to export keys while
Acra EE will use EnumerateExportedKeysByClass with a derived classifier.
The classifier may use NewExported*Key() functions to describe keys.

Provide additional utilities for exporting symmetric keys that are
stored either encrypted or as plaintext. Currently Acra CE keeps just
authentication symmetric key which is stored as plaintext. However,
Acra EE keeps many of its symmetric keys encrypted.

Since storing keys in plaintext is definitely not a norm, provide a
separate API for dealing with them, giving shorter names to the
encrypted key API.

We kind of abuse the "ID" field for symmetric keys to store encryption
context there. It's probably okay for now since we are not going to
access it directly.

* Key import API for key store v2

Add low-level method ImportKeyFileV1() which allows key store v2 to
import data exported from key store v1.

This method imports only a single key but usually you would like to
import an entire key store. It leaves transactionality to the caller.
That is, we must ensure that the key store is in consistent state
(ideally, unchanged) if importing process fails midway. It's not easy
to rollback changes so we'll introduce restrictions on importing.

Note that KeyFileImportV1 is also an interface which allows Acra EE
to override and refine importing process. Normally it would handle
extensions and then delegate keys with unknown purpose to Acra CE.

* acra-migrate-keys tool

Create a tool for migrating keys between key store formats. Currently it
supports only v1 -> v2 migration.

With default settings it is enough to simply run "acra-migrate-keys" to
migrate keys from ".acrakeys" (v1) into ".acrakeys.migrated" (v2). After
successful migration the administrator has to manually reconfigure Acra
services to use the new key store. It is possible to explicitly specify
migration details via --src... and --dst...  options if necessary.

There is a --dry_run option which performs migration in memory, without
writing results on disk. You can use it to test that everything is in
place, the keys can be read and understood.

By default the migration tool will refuse to write output into already
existing key store. This is because the migration process is not
transactional and idempotent. Migrating the same key store twice will
result in duplicate keys, and failure to migrate some keys will still
migrate other successful keys. This behavior can be overriden with
--force option, allowing the migration to proceed.

Note that master keys are expected in environment variables, as usual.
You will need to set all variables for successful migration:

  - ACRA_MASTER_KEY
  - ACRA_MASTER_ENCRYPTION_KEY
  - ACRA_MASTER_SIGNATURE_KEY

Most of the functions and data structures of the migration tool are
exported. This is an experimental attempt to reduce possible code
duplication in Acra EE migration tool. Hopefully it will be able to
reuse some of this code.

* Dry run acra-migrate-keys in configuration tests

Configuration test currently run all available tools. Make sure that
acra-migrate-keys is a) built, b) run with --dry_run and correct path
to existing key store v1 directory.

This is not a proper and final way to test it, just make sure the tests
don't break. Later we will need to be more accurate with key store
versioning and paths.

* Move zeroization utilities to "utils"

Let's have them available for common use, if necessary. They will surely
come handy some day.

* Avoid unnecessary allocations of default classifier

Use a global "constant variable" to avoid reallocating this object every
time we need it. It's a simple strategy object that does not have any
state, so we can reuse it this way.