Introduce TEST_KEYSTORE variable
Add a new configuration environment variable that can be used to control
the keystore version used for testing. By default we will be testing
with keystore v1, but you can set TEST_KEYSTORE=v2 to test with v2.

Teach test.py to pass this variable to key-generating tools as
--keystore argument. This is not done everywhere, only when we are
acting on the common shared key store. Some tests use private key
stores, with a particular version (like, for migration).