Retrieve all master key variables
Introduce a function get_master_keys() which is an extension of existing
get_master_key() that works with both v1 and v2 keys. It returns a dict
with "name => value" mapping. That way it's more useful in setUpModule()
where we configure global environment in which we will run the tests.

Basically, we setup keys for both v1 and v2, regardless of what version
we will be testing. It's easier that way.