Use acra-read-key in BaseTestCase.log
This is the last place in the test suite where we read keys directly
from the filesystem. Update the implementation to decouple it from
filesystem implementation and update call sites accordingly.

We change the logging format a bit to adapt to these changes by removing
unnecessary information. This should not cause debugging issues as all
relevant information about the keys is still present.

This method is occasionally used to read poison record keys so add this
feature to acra-read-key via "poison-public" and "poison-private" types.