Support key store v2 in acra-keymaker tests
Update the tests to explicitly set the keystore version when testing
with a temporary directory.

we need to pass proper environment variables, not only hardcoded
ACRA_MASTER_KEY. Use get_master_keys() to get a list of necessary
variables and fill them in with values that are tested.

Finally, keystore v2 will verify signatures on the keys so we cannot
run the test for long signature key in the same directory after the
standard key test. Create a new test directory for each run.