Regenerate configurations for all tools
I'm tired of getting red builds because I forgot to add a new tool into
the list in configs/regenerate.sh. Tests verify all tools present in
the "cmd" directory so this script should do the same.

Update "cmd/regenerate.sh" to regenerate configurations for all tools
found in "cmd" directory. The syntax in the same but now you don't have
to add the tools manually, only call it with "acra-..." prefix.