Run integration tests with key store v2 (#361)
* Introduce TEST_KEYSTORE variable

Add a new configuration environment variable that can be used to control
the keystore version used for testing. By default we will be testing
with keystore v1, but you can set TEST_KEYSTORE=v2 to test with v2.

Teach test.py to pass this variable to key-generating tools as
--keystore argument. This is not done everywhere, only when we are
acting on the common shared key store. Some tests use private key
stores, with a particular version (like, for migration).

* Retrieve all master key variables

Introduce a function get_master_keys() which is an extension of existing
get_master_key() that works with both v1 and v2 keys. It returns a dict
with "name => value" mapping. That way it's more useful in setUpModule()
where we configure global environment in which we will run the tests.

Basically, we setup keys for both v1 and v2, regardless of what version
we will be testing. It's easier that way.

* Read keys through acra-read-key

Introduce a new utility "acra-read-key" which can be used by tests
to read (storage) keys from the keystore. It's not intended for general
use but placed into "cmd" for technical reasons (the tests expect that
all binaries are located there, at top level).

Add a utility function read_key() that reads keys with acra-read-key
built with currently tested Go version. Use this function to implement
existing key-reading utilities like read_storage_public_key(). Note that
the tool expects master keys to be supplied through the environment so
read_storage_private_key() and read_zone_private_key() do not accept the
master key argument now.

Finally, update (almost) all places in the code where we did not use the
common utilities to read keys and instead read files directly from the
filesystem. Now they should be read with acra-read-keys in order to
handle keystore v1 & v2 transparently.

* Drop decrypt_private_key() utility

This function is no longer used now: keys are read by acra-read-keys
tool. Since the encrypted private key format is now not that simple,
this function is unlikely to be useful. Drop it.

* Drop get_master_key() utility

Preceding changes have removed last uses of this function too. It has
been completely replaced with more generic get_master_keys(). We don't
need get_master_key() anymore.

* Use acra-read-key in BaseTestCase.log

This is the last place in the test suite where we read keys directly
from the filesystem. Update the implementation to decouple it from
filesystem implementation and update call sites accordingly.

We change the logging format a bit to adapt to these changes by removing
unnecessary information. This should not cause debugging issues as all
relevant information about the keys is still present.

This method is occasionally used to read poison record keys so add this
feature to acra-read-key via "poison-public" and "poison-private" types.

* Configuration testing of acra-read-key

Allow the utility to dry-run without any useful output when key type
is not specified. However, we still need a valid key store so update
the arguments used during configuration testing to point into one.
Also, add the generated configuration as tests require it.

* Support key store v2 in acra-keymaker tests

Update the tests to explicitly set the keystore version when testing
with a temporary directory.

we need to pass proper environment variables, not only hardcoded
ACRA_MASTER_KEY. Use get_master_keys() to get a list of necessary
variables and fill them in with values that are tested.

Finally, keystore v2 will verify signatures on the keys so we cannot
run the test for long signature key in the same directory after the
standard key test. Create a new test directory for each run.

* Skip key removal tests for key store v2

Unfortunately, key store v2 does not support the hack with removing key
files from the key store to remove public or private parts of the keys.
Temporarily skip tests that require this when testing keystore v2.

These tests will be enabled back as soon as keystore v2 becomes capable
of supporting this use case.

* Regenerate configurations for all tools

I'm tired of getting red builds because I forgot to add a new tool into
the list in configs/regenerate.sh. Tests verify all tools present in
the "cmd" directory so this script should do the same.

Update "cmd/regenerate.sh" to regenerate configurations for all tools
found in "cmd" directory. The syntax in the same but now you don't have
to add the tools manually, only call it with "acra-..." prefix.

* Run key store v2 tests on CircleCI

Since CircleCI does not really support testing matrices, add a new job
that verifies PostgreSQL + key store v2. Actually, all other jobs run
fine with TEST_KEYSTORE=v2, but it's too bothersome to copy them all
and we don't really need to test them that hard right now. This may
change later when key store v2 is no longer experimental.