Log hooks (#359)
* fix parsing Parse packet from pgsql protocol (#344)

* Update Readme, fix links (#345)

* Update Readme, fix links

* Update README.md

* support rotation of acrastructs encrypted without zones (#347)

* add support of acrastructs encrypted with client-id to acra-rotate

* encryption api for acra-translator (#348)

* add encrypt method for grpc and http api

* Update golang version for docker image builds (#349)

* Fix themis dependency in the docker image (#350)

Earlier, we changed the prefix of paths where Themis installed when building from sources. Now it is `/usr/local/lib/` instead of `/usr/lib/`. But in the case when we build a new docker image from scratch which does not include system configuration files, libraries cannot be found.

Let's install Themis from packages into build image and always use `stable` branch regardless of the Acra branch.

* refactor data serialization (#346)

* remove hex/escape pg decryptor, use only binary

* Added CCPA mention to README.md (#351)

* Update README.md

* Added GDPR-related links

* Log formatting hooks

Provide extensibility hooks for log formatters which allow to inspect
and modify formatted entries.

Use logging.CustomizeLoggingWithHooks() to configure a logger with
additional hooks. Installed hooks will be executed for each entry:

  - WillFormat() is called before the entry is formatted. You may
    inspect its fields, add new extra fields, etc.

  - DidFormat() is called after the entry is formatter. The original
    entry as well as a formatter buffer are passed to the hooks. They
    may inspect the fields once again and process serialized buffer.

Hooks may return an error, in which case formatting is stopped and this
error is returned to the caller.

* sync with cop branch and ee code (#355)

sync work from acra-ee/cop branches

Co-authored-by: Alexei Lozovsky <alexei@cossacklabs.com>

* add builder for log customization

* gofmt

* add constants with format strings

* requested changes

* Don't set output to stderr explicitly

* Drop dead code

Co-authored-by: Lagovas <lagovas.lagovas@gmail.com>
Co-authored-by: vixentael <vixentael@users.noreply.github.com>
Co-authored-by: Dmytro Shapovalov <shad@shad.in.ua>
Co-authored-by: Karen Sawrey <karen@cossacklabs.com>
Co-authored-by: Alexei Lozovsky <alexei@cossacklabs.com>