fork_connector: accept keyword command-line params
Teach fork_connector() to translate additional keyword parameters into
command-line options passed to acra-connector binary, similar to how
fork_acra() does that. This is implemented by using a dictionary instead
of a list to prepare the final parameter set. It also makes it possible
to override the default parameters if necessary.