Test key store migration
Add a simple test to verify v1 -> v2 key store migration path. Since
we'll be changing the keystore layout, we cannot use global shared
key store and have to reinvent some wheels. Let's at least make it
easy to write new tests if necessary.

The tests verify that:

  - we are able to put and get data before migration

  - the migration itself completes successfully

  - the migration keeps the data unchanged and accessible

  - new data can be added after migration

We check both default and zone mode in the same test to verify that they
are not getting mixed, and that both storage and zone keys are migrated.

The tests run as a part of the test matrix, so we'll test
PostgreSQL/MySQL and TLS/Secure Session variations.