Common utils.ReadFile: acra-server
getAuthDataFromFile() does not really do much aside from providing a
marginally better error message in case the auth data file is not found.
Instead of all of this just use utils.ReadFile() which reads the file
just fine and returns an error with absolute path in it if the file
does not exist.