Relative key store path: acra-rollback
Pass the path given by the user as is to the key store, do not fix it
up to be an absolute path. While this does get us improved errors for
logging, this also messes with non-standard filesystem backends which
may not work with local filesystem paths.

The key files are stored encrypted, but full paths are not used as
context (only bare IDs), so this does not break existing deployments.