Forbid v2 master keys to be the same
Keys should not be reused for multiple purposes, that's why key store v2
uses two distinct keys: one for encrypting private key data, and one for
signing all key store data. Let's detect configuration errors and refuse
to create or open a key store if both keys are the same.

Since the key store does not have direct acccess to the keys and cannot
check for that, the only place where we can still do it is when the keys
are retrieved from the environment. This also gives us a chance to point
out to the user what exactly is wrong with the environment.