Use key store paths as specified by user (#367)
* Common utils.ReadFile: benchmarks

Just use common utils.ReadFile() instead of direct usage of
filepath.Abs() to get better logging for errors.

Extract file path constants for improved readability.

* Common utils.ReadFile: acra-server

getAuthDataFromFile() does not really do much aside from providing a
marginally better error message in case the auth data file is not found.
Instead of all of this just use utils.ReadFile() which reads the file
just fine and returns an error with absolute path in it if the file
does not exist.

* Relative key store path: acra-addzone
* Relative key store path: acra-rollback
* Relative key store path: acra-rotate

Pass the path given by the user as is to the key store, do not fix it
up to be an absolute path. While this does get us improved errors for
logging, this also messes with non-standard filesystem backends which
may not work with local filesystem paths.

The key files are stored encrypted, but full paths are not used as
context (only bare IDs), so this does not break existing deployments.