Protect key store flock with sync.Mutex (#365)
* Protect flock with sync.Mutex

BSD-style flocks are per-process. I have incorrectly assumed that these
are "Linux kernel processes" (= per-thread), however this turned out to
be incorrect. These locks are per-PID. Furthermore, they are recursive
in a weird way: if a process requests a shared lock when holding an
exclusive lock (or vice versa), the lock is transformed into the new
type in a non-atomic way (that is, the lock is released and re-grabbed).
And finally, goroutines do not map directly onto system threads, so even
if flocks did work per-thread, this would still be incorrect in Go.

Effectively, this means that we should keep track of whether we have the
lock locked and in which way. Add a sync.Mutex which will ensure that
at most one goroutine in the process keeps os.File locked.

With these changes it does not make sense to have a single Unlock() call
to unlock after both Lock() and RLock(). Introduce RUnlock() method that
mirrors sync.RWMutex.

* Avoid recursive shared locking

Test for shared locks were incorrect. Just like sync.RWMutex, we do not
allow recursive locks, even shared ones. Different goroutines can
acquire read lock concurrently, but the same goroutine cannot do so
multiple times. Doing it results in a deadlock.