Rename acra-read-key => acra-key-tool
This utility is going to be out "multitool" for key manipulation.
Initially it was thought to be useful only for testing but it seems
that it will be a good point to introduce key exporting functionality
and it may also replace acra-keymaker in the future.

However, that's in the future. Right now, rename the tool to prepare it
for extensions, rename the command-line option to be more specific, and
move key reading implementation into a separate function.